+++ collections/sets/ordered_test.go
@@ -12,6 +12,9 @@ func TestOrderedSet(t *testing.T) {
 	testOrderedSet(
 		t,
 		NewOrdered[string],
+		NewOrderedFromSeq[string],
+		NewOrderedFromKeys[string],
+		NewOrderedFromValues[string],
 		cmp.Compare[string],
 		func(m string) bool { return len(m)%2 == 0 },
 		rapid.String(),