+++ collections/sets/proto_test.go
@@ -13,6 +13,9 @@ func TestProtoSet(t *testing.T) {
 	testSet(
 		t,
 		NewProto[*ProtoStubA],
+		NewProtoFromSeq[*ProtoStubA],
+		NewProtoFromKeys[*ProtoStubA],
+		NewProtoFromValues[*ProtoStubA],
 		func(x, y *ProtoStubA) bool { return proto.Equal(x, y) },
 		func(m *ProtoStubA) bool { return len(m.Value)%2 == 0 },
 		rapid.Custom(